+++ src/main/java/com/ml/bayes/NaiveBayesModel.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * @auther: liuwenyi
+ * @author liuwenyi
  * @date 2019/5/8 13:28
  */
 @Slf4j
@@ -21,8 +21,8 @@ public class NaiveBayesModel {
         NaiveBayesModel.categoryList = categoryList;
     }
 
-    public Double predict(double[] data) {
-        Map<Double, Double> rateMap = Maps.newHashMap();
+    public Integer predict(double[] data) {
+        Map<Integer, Double> rateMap = Maps.newHashMap();
         Integer predictLen = data.length;
         for (Category category : categoryList) {
             Double probability = category.getProbability();
@@ -42,7 +42,7 @@ public class NaiveBayesModel {
             result = result * probability;
             rateMap.put(category.getIndex(), result);
         }
-        List<Map.Entry<Double, Double>> list = Lists.newArrayList(rateMap.entrySet());
+        List<Map.Entry<Integer, Double>> list = Lists.newArrayList(rateMap.entrySet());
         list.sort((o1, o2) -> (o2.getValue()).compareTo(o1.getValue()));
         return list.get(0).getKey();
     }
